+++ CHANGELOG.md
@@ -18,8 +18,8 @@
 
 ### 💥 Breaking Changes
 
-* `Column.exportFormat` and the `ExcelFormat` enum have been renamed `Column.excelFormat` and
-  `ExcelFormat` respectively.
+* `Column.exportFormat`, `Column.exportWidth` and the `ExportFormat` enum have been renamed
+ `Column.excelFormat`, `Column.excelWidth` and `ExcelFormat` respectively.
 * `Store.reuseRecords` must now be explicitly set on Stores with large datasets that wish to cache
   records by raw data identity (see above).
 * `Record` class renamed to `StoreRecord` in anticipation of upcoming changes to JavaScript standard